Adding Improved Vehicle Persistence compatibility to VCD
The Improved Vehicle Persistence, due to storing the list of unlocked player vehicles in a separate persistent array, will break VCD's ability to be aware of what vehicles are already owned, as well as the ability to sell owned vehicles. This PR adds the necessary fixes to restore functionality behind conditional compilation flags.